Fix infinite loop reading 64-bit MP4 box sizes
mp4ff_read_int64 accumulates the size out of an int8_t buffer without the
(uint8_t) cast that its 16/24/32-bit siblings in the same file all use, so
any byte above 0x7F sign-extends and sets every higher bit of the result.

parse_atoms then seeks to position + size - header_size. With a corrupted
size that sum can wrap around and land back on an already visited offset,
and since the loop only ends on a short read, the walk restarts from there
and never terminates. Upstream FAAD2 reads this field through a uint8_t
buffer; the signed one was introduced here.

+++ src/modules/synced/faad/src/mp4util.c
@@ -91,7 +91,7 @@ uint64_t mp4ff_read_int64(mp4ff_t *f) {
   mp4ff_read_data(f, data, 8);
 
   for (i = 0; i < 8; i++) {
-    result |= ((uint64_t)data[i]) << ((7 - i) * 8);
+    result |= ((uint64_t)(uint8_t)data[i]) << ((7 - i) * 8);
   }
 
   return result;
